requirements.txt
streamlit>=1.32.0
openai>=1.14.0
pdfplumber>=0.11.0
python-pptx>=0.6.23
moviepy>=1.0.3
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0
